class-e-back/requirements.txt
blinker==1.9.0
click==8.3.0
Flask==3.1.2
Flask-Cors==3.0.10
itsdangerous==2.2.0
Jinja2==3.1.6
MarkupSafe==3.0.2
Werkzeug==3.1.3
PyPDF==3.0.1
google-generativeai
python-dotenv